requirements.txt
python-dotenv
icalendar
jq
